snap/quota: move quota package here
TODO about another package placement
